# Conflicts:
#	AufträgeOrgadata/MainWindow.xaml
#	AufträgeOrgadata/MainWindow.xaml.cs
#	AufträgeOrgadata/Main_auftrag.cs
#	AufträgeOrgadata/bin/Debug/AufträgeOrgadata.exe
#	AufträgeOrgadata/bin/Debug/AufträgeOrgadata.pdb
#	AufträgeOrgadata/obj/Debug/AufträgeOrgadata.csproj.GenerateResource.Cache
#	AufträgeOrgadata/obj/Debug/AufträgeOrgadata.csprojResolveAssemblyReference.cache
#	AufträgeOrgadata/obj/Debug/AufträgeOrgadata.exe
#	AufträgeOrgadata/obj/Debug/AufträgeOrgadata.g.resources
#	AufträgeOrgadata/obj/Debug/AufträgeOrgadata.pdb
#	AufträgeOrgadata/obj/Debug/AufträgeOrgadata_MarkupCompile.cache
#	AufträgeOrgadata/obj/Debug/AufträgeOrgadata_MarkupCompile.lref
#	AufträgeOrgadata/obj/Debug/AusstattungWindow.g.i.cs
#	AufträgeOrgadata/obj/Debug/Kunde.g.i.cs
#	AufträgeOrgadata/obj/Debug/MainWindow.baml
#	AufträgeOrgadata/obj/Debug/MainWindow.g.cs
#	AufträgeOrgadata/obj/Debug/MainWindow.g.i.cs
#	AufträgeOrgadata/obj/Debug/PAddChange.g.i.cs
#	AufträgeOrgadata/obj/Debug/ProWindow.g.i.cs
#	AufträgeOrgadata/obj/Debug/Stammdaten.g.i.cs
#	AufträgeOrgadata/obj/Debug/add_edit.g.i.cs
#	AufträgeOrgadata/obj/Debug/database.g.i.cs

+++ AufträgeOrgadata/Stammdaten.xaml.cs
@@ -1,40 +1,27 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
 using MySql.Data.MySqlClient;
 
 namespace AufträgeOrgadata
 {
-    /// <summary>
-    /// Interaktionslogik für Stammdaten.xaml
-    /// </summary>
-    public partial class Stammdaten : Window
+    public partial class Stammdaten
     {
         public Stammdaten()
         {
             InitializeComponent();
         }
+
         public class TStammDaten
         {
             public int ID { get; set; }
-            public String StammName { get; set; }
+            public string StammName { get; set; }
         }
+
         private void MenuItem_Click(object sender, RoutedEventArgs e)
         {
-            this.Close();
-
-        }  
-
+            Close();
+        }
 
         public class StammdatenCs
         {
@@ -50,31 +37,32 @@ namespace AufträgeOrgadata
             {
                 login lgn = new login();
 
-                string uid, pw, server, port, db, table;
-                uid = lgn.lgnList[0].uid;
-                pw = lgn.lgnList[0].pw;
-                server = lgn.lgnList[0].server;
-                port = lgn.lgnList[0].port;
-                db = lgn.lgnList[0].db;
-                table = lgn.lgnList[0].table;
+                var uid = lgn.lgnList[0].uid;
+                var pw = lgn.lgnList[0].pw;
+                var server = lgn.lgnList[0].server;
+                var port = lgn.lgnList[0].port;
+                var db = lgn.lgnList[0].db;
+                var table = lgn.lgnList[0].table;
 
-                String connstring = "uid=" + uid + ";" + "password=" + pw + ";" + "server=" + server + ";" + "port=" + port + ";" + "database=" + db + ";" + "table=" + table + ";";
+                var connstring = "uid=" + uid + ";" + "password=" + pw + ";" + "server=" + server + ";" + "port=" + port +
+                                 ";" + "database=" + db + ";" + "table=" + table + ";";
                 MySqlConnection conn = new MySqlConnection(connstring);
 
                 try
                 {
                     conn.Open();
-                    MySqlCommand cmd = new MySqlCommand("SELECT * FROM stammdaten");
-                    cmd.Connection = conn;
+                    MySqlCommand cmd = new MySqlCommand("SELECT * FROM stammdaten") {Connection = conn};
 
                     using (MySqlDataReader Reader = cmd.ExecuteReader())
                     {
                         while (Reader.Read())
                         {
 
-                            TStammDaten StammN = new TStammDaten();
-                            StammN.ID = int.Parse(Reader["ID"].ToString());
-                            StammN.StammName = Reader["StammName"].ToString();
+                            TStammDaten StammN = new TStammDaten
+                            {
+                                ID = int.Parse(Reader["ID"].ToString()),
+                                StammName = Reader["StammName"].ToString()
+                            };
                             StammDatenliste.Add(StammN);
                         }
                     }
@@ -92,12 +80,12 @@ namespace AufträgeOrgadata
         {
             StammdatenCs SName = new StammdatenCs();
 
-            for (int i = 0; i < SName.StammDatenliste.Count; i++)
+            foreach (TStammDaten t in SName.StammDatenliste)
             {
                 lvStammDaten.Items.Add(new TStammDaten
                 {
-                    ID = SName.StammDatenliste[i].ID,
-                    StammName = SName.StammDatenliste[i].StammName
+                    ID = t.ID,
+                    StammName = t.StammName
 
                 });
             }
@@ -105,50 +93,31 @@ namespace AufträgeOrgadata
 
         public void mDelete_Click(object sender, RoutedEventArgs e)
         {
-
+            if (lvStammDaten.SelectedIndex < 0) return;
             var selectitem = (dynamic)lvStammDaten.SelectedItems[0];
+            var connstring = "Server = localhost; database = auftraege; uid = root ";
 
-            String connstring = "Server = localhost; database = auftraege; uid = root ";
-
-            MySqlConnection conn = new MySqlConnection(connstring);
-
-
-            try
+            using (MySqlConnection conn = new MySqlConnection(connstring))
             {
-                conn.Open();
+                try
+                {
+                    conn.Open();
+                    var cmd = new MySqlCommand("Delete from  stammdaten where ID = ?ItemClick");
 
-                MySqlCommand cmd = new MySqlCommand("Delete from  stammdaten where ID = ?ItemClick");
-  
+                    cmd.Parameters.AddWithValue("?ItemClick", selectitem.StammName);
 
-                cmd.Parameters.AddWithValue("?ItemClick", selectitem.StammName);
+                    MessageBox.Show(Convert.ToString(selectitem.ID));
 
-                MessageBox.Show(Convert.ToString(selectitem.ID));
+                    cmd.Connection = conn;
+                    cmd.ExecuteNonQuery();
 
-                
-                cmd.Connection = conn;
-                cmd.ExecuteNonQuery();
-
-                conn.Close();
+                    conn.Close();
+                }
+                catch (Exception e1)
+                {
+                    MessageBox.Show(e1.Message);
+                }
             }
-            catch (Exception e1)
-            {
-                MessageBox.Show(e1.Message);
-            }
-
-
         }
     }
-    }
-
-
-
-
-
-
-
-
-
-
-
-
-
+}
